fix: refactor according to suggestions
